requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.30.0
backports.asyncio.runner==1.2.0
bcrypt==4.0.1
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
coverage==7.10.7
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.117.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.43
starlette==0.48.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
WTForms==3.1.2
